Data.Time.Calendar.YearMonth

Description

This module exports a type representing months in the Gregorian calendar (YearMonth) and associated functions.

Documentation

newtype YearMonth

Type representing months in the Gregorian calander.

Constructors

YearMonth (Integer, Int)

toYearMonth :: Day -> YearMonth

Returns the specific month in which the given date falls.

fromYearMonth

Arguments

:: Int
The day of the month. The result will be clamped to the range from 1 to the last day of the month. Use fromYearMonth 31 for the last day of every month.

-> YearMonth

-> Day

Returns a date for the specifc month, given the day of the month.