Web.UK.Police

Description

This module exports types and functions used to implement bindings to the UK Home Office's Police API.

The following is an example application using the API.



module Main where

import Data.Either (rights)
import System.Environment (getArgs)

import Data.Text (pack)
import Data.Text.Read (double)
import Network.HTTP.Client (newManager)
import Network.HTTP.Client.TLS (tlsManagerSettings)

import Web.UK.Police

main :: IO ()
main = do
  args <- getArgs
  let ll = CrimeLatLng $ coord args
  mgr <- newManager tlsManagerSettings
  crimesStreetDates mgr >>= \case
    Right result -> do
      let ms = availableMonths result
      clrs <- rights <$> mapM (crimesAtLocation mgr ll . Just) ms
      let css = concatMap (\(CrimesAtLocationResponse cs) -> cs) clrs
          crimes = map (\(CrimesStreet crime _) -> crime) css
          ccs = map (\c -> ( cMonth c
                           , cCategory c
                           , streetName . slStreet <$> cLocation c)) crimes
      mapM_ print ccs
    _ -> pure ()
 where
  coord args = let coord' n = either undefined fst $ double $ pack $ args!!n
               in  LatLng (coord' 0) (coord' 1)

API functions

Availability-related

crimesStreetDates :: Manager -> IO (Either ClientError DataAvailabilityResponse)

Returns a list of available data sets.

crimeLastUpdated :: Manager -> IO (Either ClientError CrimeLastUpdatedResponse)

When the crime data was last updated.

Force-related

forces :: Manager -> IO (Either ClientError ForcesResponse)

A list of all the police forces available via the API.

force

Arguments

:: Manager

-> TForce
Territorial police force.

-> IO (Either ClientError ForceResponse)

Returns information about a specific police force.

seniorOfficers

Arguments

:: Manager

-> TForce
Territorial police force.

-> IO (Either ClientError SeniorOfficersResponse)

Returns a list of the senior officers for a given police force.

Neighbourhood-related

neighbourhoods

Arguments

:: Manager

-> TForce
Territorial police force.

-> IO (Either ClientError NeighbourhoodsResponse)

Returns a list of valid neighbourhood identifiers and names for a given police force.

neighbourhood

Arguments

:: Manager

-> TForce
Territorial police force.

-> NeighbourhoodId

-> IO (Either ClientError NeighbourhoodResponse)

Returns a description of a given neighbourhood.

boundary

Arguments

:: Manager

-> TForce
Territorial police force.

-> NeighbourhoodId

-> IO (Either ClientError BoundaryResponse)

Returns a list of latitude/longitude pairs that make up the boundary of a given neigbourhood.

people

Arguments

:: Manager

-> TForce
Territorial police force.

-> NeighbourhoodId

-> IO (Either ClientError PeopleResponse)

Returns a list of members of team for a given neighbourhood.

events

Arguments

:: Manager

-> TForce
Territorial police force.

-> NeighbourhoodId

-> IO (Either ClientError EventsResponse)

Returns a list of future events for a given neighbourhood.

priorities

Arguments

:: Manager

-> TForce
Territorial police force.

-> NeighbourhoodId

-> IO (Either ClientError PrioritiesResponse)

Returns a list of priorities for a given neighbourhood.

locateNeighbourhood :: Manager -> LatLng -> IO (Either ClientError LocateNeighbourhoodResponse)

Find the neighbourhood policing team responsible for a particular area.

Crime reports-related

crimesCategories

Arguments

:: Manager

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError CrimesCategoriesResponse)

Returns a list of valid categories for a given data set date.

crimesStreet

Arguments

:: Manager

-> CrimesCategory

-> CrimeArea

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError CrimesStreetResponse)

Crimes at street-level; either within a 1 mile radius of a single point, or within a custom area.

crimesAtLocation

Arguments

:: Manager

-> CrimeLocation

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError CrimesAtLocationResponse)

Returns just the crimes which occurred at the specified location, rather than those within a radius. If given latitude and longitude, finds the nearest pre-defined location and returns the crimes which occurred there.

crimesNoLocation :: Manager -> CrimesCategory -> Force -> Maybe YearMonth -> IO (Either ClientError CrimesNoLocationResponse)

Returns a list of crimes that could not be mapped to a location.

outcomesStreet

Arguments

:: Manager

-> LocationId

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError OutcomesAtLocationResponse)

Outcomes at street-level, at a specific location.

outcomesAtLocation

Arguments

:: Manager

-> CrimeArea

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError OutcomesAtLocationResponse)

Outcomes at street-level; either within a 1 mile radius of a single point, or within a custom area.

outcomesForCrime :: Manager -> PersistentId -> IO (Either ClientError OutcomesForCrimeResponse)

Returns the outcomes (case history) for the specified crime.

Stop and search-related

stopsStreet

Arguments

:: Manager

-> CrimeArea

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError StopsStreetResponse)

Stop and searches at street-level; either within a 1 mile radius of a single point, or within a custom area.

stopsAtLocation

Arguments

:: Manager

-> LocationId

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError StopsAtLocationResponse)

Stop and searches at a particular location.

stopsNoLocation

Arguments

:: Manager

-> Force
Police force.

-> Maybe YearMonth
The lastest month will be used by default.

-> IO (Either ClientError StopsNoLocationResponse)

Stop and searches that could not be mapped to a location.

stopsForce

Arguments

:: Manager

-> Force
Police force.

-> Maybe YearMonth
The latest month will be used by default.

-> IO (Either ClientError StopsForceResponse)

Stop and searches reported by a particular force.

Utilities

availableMonths :: DataAvailabilityResponse -> [YearMonth]

Returns a sorted list of months for which data is available, given a DataAvailabilityResponse.

availableStopAndSearch :: DataAvailabilityResponse -> Force -> [YearMonth]

Returns a sorted list of months for which stop and search data is available, given a DataAvailabilityResponse and a participating police force.

tForce :: Force -> Maybe TForce

Attempts to cast a police force to a territorial police force.

API

type UKPoliceAPI = ("crimes-street-dates" :> Get '[JSON] DataAvailabilityResponse) :<|> (("forces" :> Get '[JSON] ForcesResponse) :<|> (("forces" :> (Capture "force" TForce :> Get '[JSON] ForceResponse)) :<|> (("forces" :> (Capture "force" TForce :> ("people" :> Get '[JSON] SeniorOfficersResponse))) :<|> (("crimes-street" :> (Capture "crime-category" CrimesCategory :> (QueryParam "lat" Double :> (QueryParam "lng" Double :> (QueryParam "poly" Poly :> (QueryParam "date" YearMonth :> Get '[JSON] CrimesStreetResponse)))))) :<|> (("outcomes-at-location" :> (QueryParam "location_id" LocationId :> (QueryParam "lat" Double :> (QueryParam "lng" Double :> (QueryParam "poly" Poly :> (QueryParam "date" YearMonth :> Get '[JSON] OutcomesAtLocationResponse)))))) :<|> (("crimes-at-location" :> (QueryParam "location_id" LocationId :> (QueryParam "lat" Double :> (QueryParam "lng" Double :> (QueryParam "date" YearMonth :> Get '[JSON] CrimesAtLocationResponse))))) :<|> (("crimes-no-location" :> (QueryParam "category" CrimesCategory :> (QueryParam "force" Force :> (QueryParam "date" YearMonth :> Get '[JSON] CrimesNoLocationResponse)))) :<|> (("crime-categories" :> (QueryParam "date" YearMonth :> Get '[JSON] CrimesCategoriesResponse)) :<|> (("crime-last-updated" :> Get '[JSON] CrimeLastUpdatedResponse) :<|> (("outcomes-for-crime" :> (Capture "peristent_id" PersistentId :> Get '[JSON] OutcomesForCrimeResponse)) :<|> ((Capture "force" TForce :> ("neighbourhoods" :> Get '[JSON] NeighbourhoodsResponse)) :<|> ((Capture "force" TForce :> (Capture "neighbourhood_id" NeighbourhoodId :> Get '[JSON] NeighbourhoodResponse)) :<|> ((Capture "force" TForce :> (Capture "neighbourhood_id" NeighbourhoodId :> ("boundary" :> Get '[JSON] BoundaryResponse))) :<|> ((Capture "force" TForce :> (Capture "neighbourhood_id" NeighbourhoodId :> ("people" :> Get '[JSON] PeopleResponse))) :<|> ((Capture "force" TForce :> (Capture "neighbourhood_id" NeighbourhoodId :> ("events" :> Get '[JSON] EventsResponse))) :<|> ((Capture "force" TForce :> (Capture "neighbourhood_id" NeighbourhoodId :> ("priorities" :> Get '[JSON] PrioritiesResponse))) :<|> (("locate-neighbourhood" :> (QueryParam "q" LatLng :> Get '[JSON] LocateNeighbourhoodResponse)) :<|> (("stops-street" :> (QueryParam "lat" Double :> (QueryParam "lng" Double :> (QueryParam "poly" Poly :> (QueryParam "date" YearMonth :> Get '[JSON] StopsStreetResponse))))) :<|> (("stops-at-location" :> (QueryParam "location_id" LocationId :> (QueryParam "date" YearMonth :> Get '[JSON] StopsAtLocationResponse))) :<|> (("stops-no-location" :> (QueryParam "force" Force :> (QueryParam "date" YearMonth :> Get '[JSON] StopsNoLocationResponse))) :<|> ("stops-force" :> (QueryParam "force" Force :> (QueryParam "date" YearMonth :> Get '[JSON] StopsForceResponse)))))))))))))))))))))))

Type representing the UK Home Office's Police API.

api :: Proxy UKPoliceAPI

The api.

End point responses

Availability-related

newtype DataAvailabilityResponse

Type representing the response to an end point of the API.

Constructors

DataAvailabilityResponse [DataAvailability]

newtype CrimeLastUpdatedResponse

Type representing the response to an end point of the API.

Constructors

CrimeLastUpdatedResponse CrimeLastUpdated

Force-related

newtype ForcesResponse

Type representing the response to an end point of the API.

Constructors

ForcesResponse [ForceId]

newtype ForceResponse

Type representing the response to an end point of the API.

Constructors

ForceResponse ForceDescription

newtype SeniorOfficersResponse

Type representing the response to an end point of the API.

Constructors

SeniorOfficersResponse [Person]

Neighbourhood-related

newtype NeighbourhoodsResponse

Type representing the response to an end point of the API.

Constructors

NeighbourhoodsResponse [Neighbourhood]

data NeighbourhoodResponse

Type representing the response to an end point of the API.

Constructors

NeighbourhoodResponse Neighbourhood NeighbourhoodDescription (Maybe ContactDetails)

newtype BoundaryResponse

Type representing the response to an end point of the API.

Constructors

BoundaryResponse [LatLng]

newtype PeopleResponse

Type representing the response to an end point of the API.

Constructors

PeopleResponse [Person]

newtype EventsResponse

Type representing the response to an end point of the API.

Constructors

EventsResponse [Event]

newtype PrioritiesResponse

Type representing the response to an end point of the API.

Constructors

PrioritiesResponse [Priority]

data LocateNeighbourhoodResponse

Type representing the response to an end point of the API.

Constructors

LocateNeighbourhoodResponse Force NeighbourhoodId

Crime reports-related

newtype CrimesCategoriesResponse

Type representing the response to an end point of the API.

Constructors

CrimesCategoriesResponse [CrimesCategoryDescription]

newtype CrimesStreetResponse

Type representing the response to an end point of the API.

Constructors

CrimesStreetResponse [CrimesStreet]

newtype CrimesAtLocationResponse

Type representing the response to an end point of the API.

Constructors

CrimesAtLocationResponse [CrimesStreet]

newtype CrimesNoLocationResponse

Type representing the response to an end point of the API.

Constructors

CrimesNoLocationResponse [CrimesStreet]

newtype OutcomesAtLocationResponse

Type representing the response to an end point of the API.

Constructors

OutcomesAtLocationResponse [OutcomeAtLocation]

data OutcomesForCrimeResponse

Type representing the response to an end point of the API.

Constructors

OutcomesForCrimeResponse Crime [Outcome]

Stop and search-related

newtype StopsStreetResponse

Type representing the response to an end point of the API.

Constructors

StopsStreetResponse [Stop]

newtype StopsAtLocationResponse

Type representing the response to an end point of the API.

Constructors

StopsAtLocationResponse [Stop]

newtype StopsNoLocationResponse

Type representing the response to an end point of the API.

Constructors

StopsNoLocationResponse [Stop]

newtype StopsForceResponse

Type representing the response to an end point of the API.

Constructors

StopsForceResponse [Stop]

Other types

Time and date-related

newtype YearMonth

Type representing months in the Gregorian calander.

Constructors

YearMonth (Integer, Int)

Georgraphy-related

data CrimeArea

Type representing areas used to obtain crime reports

Constructors

CrimeCentre LatLng
Centre of a circle of radius 1 mile

CrimePoly Poly
Custom area

data CrimeLocation

Type representing location used to obtain crime reports

Constructors

CrimeLocation LocationId

CrimeLatLng LatLng

data LatLng

Type representing latitude and longitude coordinates

Constructors

LatLng

Fields

lat :: Double

lng :: Double

newtype Poly

Type representing boundaries of regions

Constructors

Poly [LatLng]

newtype LocationId

Type representing identifiers for anonymised locations

Constructors

LocationId Int

data LocationType

Type representing types of locations.

Constructors

LTForce

LTBTP

UnknownLocationType Text

data Street

Type representing streets

Constructors

Street

Fields

streetId :: LocationId

streetName :: Text

data StreetLocation

Type representating streets and their locations

Constructors

StreetLocation

Fields

slLatLng :: LatLng

slStreet :: Street

data ForceLocation

Type representing locations of police.

Constructors

ForceLocation

Fields

flName :: Text

flType :: Text

flDescription :: Maybe Text

flLocation :: Maybe LatLng

flAddress :: Text

flPostcode :: Text

Force and neighbourhood-related

data Force

Type representing participating UK police forces

Constructors

TF TForce
Territorial police force

BTP
British Transport Police

data TForce

Type representing participating UK territorial police forces

Constructors

EW EWTForce
Territorial police force in England and Wales

NorthernIreland
Police Service of Northern Ireland

data EWTForce

Type representing territorial police forces in England and Wales. There is a nullary data constructor for each of the 43 territorial police forces.

Constructors

AvonAndSomerset

Bedfordshire

Cambridgeshire

Cheshire

CityOfLondon

Cleveland

Cumbria

Derbyshire

DevonAndCornwall

Dorset

Durham

DyfedPowys

Essex

Gloucestershire

GreaterManchester

Gwent

Hampshire

Hertfordshire

Humberside

Kent

Lancashire

Leicestershire

Lincolnshire

Merseyside

Metropolitan

Norfolk

NorthWales

NorthYorkshire

Northamptonshire

Northumbria

Nottinghamshire

SouthWales

SouthYorkshire

Staffordshire

Suffolk

Surrey

Sussex

ThamesValley

Warwickshire

WestMercia

WestMidlands

WestYorkshire

Wiltshire

data ForceId

Type representing identifiers and associated names of police forces.

Constructors

ForceId

Fields

forceId :: Force

forceName :: Text

data ForceDescription

Type representing descriptions of police forces.

Constructors

ForceDescription

Fields

fdId :: Force

fdName :: Text

fdDescription :: Maybe Text

fdUrl :: Text

fdTelephone :: Text

fdEngagementMethods :: Maybe [EngagementMethod]

data Neighbourhood

Type representing identification of neighbourhoods.

Constructors

Neighbourhood

Fields

nId :: NeighbourhoodId

nName :: Text

data NeighbourhoodDescription

Type representing descriptions of a neighbourhood.

Constructors

NeighbourhoodDescription

Fields

nUrlForce :: Text

nWelcomeMessage :: Maybe Text

nLinks :: [Link]

nCentre :: LatLng

nLocations :: [ForceLocation]

nPopulation :: Text

nDescription :: Text

newtype NeighbourhoodId

Type representing identifiers for neighbourhoods.

Constructors

NeighbourhoodId Text

data Priority

Type representing neighbourhood priorities.

Constructors

Priority

Fields

pIssue :: Text

pIssueDate :: Maybe LocalTime

pAction :: Maybe Text

pActionDate :: Maybe LocalTime

data Person

Type representing people, either senior officers or in a neighbourhood team.

Constructors

Person

Fields

pName :: Text

pRank :: Maybe Text

pBio :: Maybe Text

pContactDetails :: Maybe ContactDetails

Contact-related

data ContactDetails

Type representing contact details.

Constructors

ContactDetails

Fields

cdEmail :: Maybe Text

cdTelephone :: Maybe Text

cdMobile :: Maybe Text

cdFax :: Maybe Text

cdWeb :: Maybe Text

cdAddress :: Maybe Text

cdFacebook :: Maybe Text

cdTwitter :: Maybe Text

cdYoutube :: Maybe Text

cdMyspace :: Maybe Text

cdBebo :: Maybe Text

cdFlickr :: Maybe Text

cdGooglePlus :: Maybe Text

cdForum :: Maybe Text

cdEMessaging :: Maybe Text

cdBlog :: Maybe Text

cdRss :: Maybe Text

data EngagementMethod

Type representing engagement methods.

Constructors

EngagementMethod

Fields

emUrl :: Text

emDescription :: Text

emTitle :: Text

data Link

Type representing links to information on the internet.

Constructors

Link

Fields

linkUrl :: Text

linkTitle :: Text

linkDescription :: Maybe Text

data Event

Type representing neighbourhood events.

Constructors

Event

Fields

eTitle :: Text

eType :: Text

eDescription :: Maybe Text

eContactDetails :: Maybe ContactDetails

eAddress :: Text

eStartDate :: LocalTime

eEndDate :: LocalTime

Crime report-related

data DataAvailability

Type representing months for which reported crime is available and a list of police forces providing stop and search information for the month in question.

Constructors

DataAvailability

Fields

dataDate :: YearMonth

dataStopAndSearch :: [Force]

data CrimesCategory

Type representing crimes categories.

Constructors

AllCrime

AntiSocialBehaviour

BicycleTheft

Burglary

CriminalDamageArson

Drugs

OtherTheft

PossessionOfWeapons

PublicOrder

Robbery

Shoplifting

TheftFromThePerson

VehicleCrime

ViolentCrime

ViolenceAndSexualOffences

OtherCrime

UnknownCrimesCategory Text

data CrimesCategoryDescription

Type representing identifiers and associated descriptions of crimes categories.

Constructors

CrimesCategoryDescription

Fields

ccUrl :: CrimesCategory

ccName :: Text

data Crime

Type representing crime reports.

Constructors

Crime

Fields

cCategory :: CrimesCategory

cLocationType :: Maybe LocationType

cLocationSubtype :: Maybe Text

cLocation :: Maybe StreetLocation

cContext :: Text

cPersistentId :: Maybe PersistentId

cId :: Integer

cMonth :: YearMonth

newtype PersistentId

Type representing persistent identifiers for UK crime reports.

Constructors

PersistentId (Vector 64 Char)

newtype CrimeLastUpdated

Type representing dates on which reported crime data was last updated.

Constructors

CrimeLastUpdated

Fields

cluDate :: Day

data CrimesStreet

Type representing reported crimes and their associated outcomes.

Constructors

CrimesStreet Crime (Maybe Outcome)

data Outcome

Type representing outcomes for crime reports.

Constructors

Outcome

Fields

oCategory :: OutcomeCategory

oPersonId :: Maybe Int

oMonth :: YearMonth

data OutcomeAtLocation

Type representing outcomes and associated crimes with a location.

Constructors

OutcomeAtLocation Crime Outcome

Stop and search-related

data Stop

Type representing stops and searches

Constructors

Stop

Fields

ssType :: StopType

ssRemovalOfMoreThanOuterClothing :: Maybe Bool

ssLocation :: Maybe StreetLocation

ssDatetime :: ZonedTime

ssGender :: Maybe Gender

ssAgeRange :: Maybe StopAgeRange

ssSelfDefinedEthnicity :: Maybe Text

ssOfficerDefinedEthnicity :: Maybe Text

ssLegislation :: Maybe Text

ssObjectOfSearch :: Maybe Text

ssOperation :: Maybe Bool

ssOperationName :: Maybe Text

ssOutcome :: Maybe StopOutcome

ssOutcomeId :: Maybe StopOutcomeId

ssOutcomeLinkedToObjectOfSearch :: Maybe Bool

data StopType

Type representing the types of stops and searches.

Constructors

PersonSearch

VehicleSearch

PersonAndVehicleSearch

data Gender

Type representing reported genders of people stopped and searched.

Constructors

Male

Female

data StopAgeRange

Type representing reported age ranges of people stopped and searched.

Constructors

Years10To17

Years18to24

Years25to34

Years35Plus

data StopOutcome

Type representing stop outcomes.

Constructors

NoStopOutcome

NothingFound

StopOutcome Text

newtype StopOutcomeId

Type representing stop and search outcomes

Constructors

StopOutcomeId Text