Web.UK.Police.Types.EWTForce

Description

This module exports a type representing territorial police forces in England and Wales (EWTForce).

Documentation

data EWTForce

Type representing territorial police forces in England and Wales. There is a nullary data constructor for each of the 43 territorial police forces.

Constructors

AvonAndSomerset

Bedfordshire

Cambridgeshire

Cheshire

CityOfLondon

Cleveland

Cumbria

Derbyshire

DevonAndCornwall

Dorset

Durham

DyfedPowys

Essex

Gloucestershire

GreaterManchester

Gwent

Hampshire

Hertfordshire

Humberside

Kent

Lancashire

Leicestershire

Lincolnshire

Merseyside

Metropolitan

Norfolk

NorthWales

NorthYorkshire

Northamptonshire

Northumbria

Nottinghamshire

SouthWales

SouthYorkshire

Staffordshire

Suffolk

Surrey

Sussex

ThamesValley

Warwickshire

WestMercia

WestMidlands

WestYorkshire

Wiltshire