Web.UK.Police.Types.OutcomeCategory

Description

This module exports a type used to represent categories for the outcome of crime reports used by the UK Home Office's Police API (OutcomeCategory).

Documentation

data OutcomeCategory

Type representing categories of outcomes for reported crimes.

Constructors

AwaitingCourtResult

CourtResultUnavailable

UnableToProceed

LocalResolution

NoFurtherAction

DeprivedOfProperty

Fined

AbsoluteDischarge

Cautioned

DrugsPossessionWarning

PenaltyNoticeIssued

CommunityPenalty

ConditionalDischarge

SuspendedSentence

Imprisoned

OtherCourtDisposal

Compensation

SentencedInAnotherCase

Charged

NotGuilty

SentToCrownCourt

UnableToProsecute

FormalActionNotInPublicInterest

ActionTakenByAnotherOrganisation

FurtherInvestigationNotInPublicInterest

FurtherActionNotInPublicInterest

UnderInvestigation

StatusUpdateUnavailable